+++ tests/testthat/test-bound_record.R
@@ -1,0 +1,6 @@
+with_mock_dir("httptest2/bndrcd", {
+  test_that("cong_bound_record works", {
+    x <- cong_bound_record()
+    expect_s3_class(x, 'tbl_df')
+  })
+})